Fixes the Delete and Update webhook dialogs (#1350)
* Broken down the webhook table pages in smaller more manageable components

* Reorganised folders and files names for Webhooks

* Refacrtored the page Webhook with ID

* Showing the Create and Delete dialogs in front of the content

* Fixed the missing map key

* Fixed the Events Select to work on all browsers and not to overlap

* Changed the border color for webhood details

* Cutting the webhook name and URL when too long in the table

* Renamed the popups to make them more intuitive

* Saving the secret when creating a new webhook
