+++ packages/www/components/Dashboard/WebhooksTable/index.tsx
@@ -1,43 +1,16 @@
 import { useCallback, useMemo, useState } from "react";
 import { useApi } from "../../../hooks";
 import Table, { Fetcher, useTableState } from "components/Dashboard/Table";
-import TextCell, { TextCellProps } from "components/Dashboard/Table/cells/text";
-import DateCell, { DateCellProps } from "components/Dashboard/Table/cells/date";
-import StatusBadge, { Variant as StatusVariant } from "../StatusBadge";
-import { dateSort, stringSort } from "components/Dashboard/Table/sorts";
-import { SortTypeArgs } from "components/Dashboard/Table/types";
-import { Column } from "react-table";
-import {
-  Box,
-  Flex,
-  Heading,
-  Link as A,
-  AlertDialog,
-  AlertDialogTitle,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogCancel,
-  AlertDialogAction,
-  Button,
-  Text,
-  useSnackbar,
-} from "@livepeer/design-system";
+import { Box, Heading } from "@livepeer/design-system";
 import { useToggleState } from "hooks/use-toggle-state";
 import { Cross1Icon, PlusIcon } from "@radix-ui/react-icons";
-import Spinner from "components/Dashboard/Spinner";
-import WebhookDialog, { Action } from "components/Dashboard/WebhookDialog";
+import CreateDialog, {
+  Action,
+} from "@components/Dashboard/WebhookDialogs/CreateEditDialog";
 import { useRouter } from "next/router";
-import { Webhook } from "@livepeer.studio/api";
-
-type WebhooksTableData = {
-  name: TextCellProps;
-  url: TextCellProps;
-  created: DateCellProps;
-  status: TextCellProps;
-};
-
-// 1 hour
-const WARNING_TIMEFRAME = 1000 * 60 * 60;
+import { makeColumns, rowsPageFromState, WebhooksTableData } from "./helpers";
+import EmptyState from "./EmptyState";
+import DeleteDialog from "./DeleteDialog";
 
 const WebhooksTable = ({ title = "Webhooks" }: { title?: string }) => {
   const router = useRouter();
@@ -45,126 +18,15 @@ const WebhooksTable = ({ title = "Webhooks" }: { title?: string }) => {
     useApi();
   const deleteDialogState = useToggleState();
   const [savingDeleteDialog, setSavingDeleteDialog] = useState(false);
-  const [openSnackbar] = useSnackbar();
   const createDialogState = useToggleState();
   const { state, stateSetter } = useTableState<WebhooksTableData>({
     tableId: "webhooksTable",
   });
 
-  const columns = useMemo(
-    () => [
-      {
-        Header: "URL",
-        accessor: "url",
-        Cell: TextCell,
-        sortType: (...params: SortTypeArgs) =>
-          stringSort("original.url.value", ...params),
-      },
-      {
-        Header: "Name",
-        accessor: "name",
-        Cell: TextCell,
-        sortType: (...params: SortTypeArgs) =>
-          stringSort("original.name.children", ...params),
-      },
-      {
-        Header: "Created at",
-        accessor: "created",
-        Cell: DateCell,
-        sortType: (...params: SortTypeArgs) =>
-          dateSort("original.created.date", ...params),
-      },
-      {
-        Header: "Status",
-        accessor: "status",
-        Cell: TextCell,
-        disableSortBy: true,
-      },
-    ],
-    []
-  );
+  const columns = useMemo(makeColumns, []);
 
   const fetcher: Fetcher<WebhooksTableData> = useCallback(
-    async (state) => {
-      const [webhooks, nextCursor, _res, count] = await getWebhooks(
-        false,
-        false,
-        state.order,
-        null,
-        state.pageSize,
-        state.cursor,
-        true
-      );
-
-      return {
-        nextCursor,
-        count,
-        rows: webhooks.map((webhook: Webhook) => {
-          return {
-            id: webhook.id,
-            name: {
-              children: webhook.name,
-              href: `/dashboard/developers/webhooks/${webhook.id}`,
-              css: {
-                cursor: "pointer",
-              },
-            },
-            url: {
-              value: webhook.url,
-              children: (
-                <A as="div" variant="primary">
-                  {webhook.url}
-                </A>
-              ),
-              href: `/dashboard/developers/webhooks/${webhook.id}`,
-              css: {
-                cursor: "pointer",
-              },
-            },
-            created: {
-              date: new Date(webhook.createdAt),
-              fallback: <i>unseen</i>,
-              href: `/dashboard/developers/webhooks/${webhook.id}`,
-              css: {
-                cursor: "pointer",
-              },
-            },
-            status: {
-              children: (
-                <Box>
-                  {!webhook.status ? (
-                    <StatusBadge
-                      variant={StatusVariant.Idle}
-                      tooltipText="No triggers yet"
-                    />
-                  ) : (webhook.status.lastFailure &&
-                      +Date.now() - webhook.status.lastFailure?.timestamp <
-                        WARNING_TIMEFRAME) ||
-                    webhook.status.lastFailure?.timestamp >=
-                      webhook.status.lastTriggeredAt ? (
-                    <StatusBadge
-                      variant={StatusVariant.Unhealthy}
-                      timestamp={webhook.status.lastFailure.timestamp}
-                      tooltipText="Last failure"
-                    />
-                  ) : (
-                    <StatusBadge
-                      variant={StatusVariant.Healthy}
-                      timestamp={webhook.status.lastTriggeredAt}
-                      tooltipText="Last triggered"
-                    />
-                  )}
-                </Box>
-              ),
-              href: `/dashboard/developers/webhooks/${webhook.id}`,
-              css: {
-                cursor: "pointer",
-              },
-            },
-          };
-        }),
-      };
-    },
+    async (state) => rowsPageFromState(state, getWebhooks),
     [getWebhooks, user.id]
   );
 
@@ -186,50 +48,19 @@ const WebhooksTable = ({ title = "Webhooks" }: { title?: string }) => {
     state.invalidate,
   ]);
 
-  const emptyState = (
-    <Flex
-      direction="column"
-      justify="center"
-      css={{
-        margin: "0 auto",
-        height: "calc(100vh - 400px)",
-        maxWidth: 450,
-      }}>
-      <Heading css={{ fontWeight: 500, mb: "$3" }}>
-        Create your first webhook
-      </Heading>
-      <Text variant="gray" css={{ lineHeight: 1.5, mb: "$3" }}>
-        Listen for events on your Livepeer Studio account so your integration
-        can automatically trigger reactions.
-      </Text>
-      <Button
-        onClick={() => createDialogState.onOn()}
-        css={{ alignSelf: "flex-start" }}
-        size="2"
-        variant="primary">
-        <PlusIcon />{" "}
-        <Box as="span" css={{ ml: "$2" }}>
-          Create webhook
-        </Box>
-      </Button>
-    </Flex>
-  );
-
   return (
-    <>
+    <Box css={{ p: "$6", mb: "$8" }}>
       <Table
         columns={columns}
         fetcher={fetcher}
         state={state}
         stateSetter={stateSetter}
         rowSelection="all"
-        emptyState={emptyState}
+        emptyState={<EmptyState createDialogState={createDialogState} />}
         header={
-          <>
-            <Heading size="2" css={{ fontWeight: 600 }}>
-              {title}
-            </Heading>
-          </>
+          <Heading size="2" css={{ fontWeight: 600 }}>
+            {title}
+          </Heading>
         }
         selectAction={{
           onClick: deleteDialogState.onOn,
@@ -256,84 +87,24 @@ const WebhooksTable = ({ title = "Webhooks" }: { title?: string }) => {
         }}
       />
 
-      <AlertDialog
-        open={deleteDialogState.on}
-        onOpenChange={deleteDialogState.onOff}>
-        <AlertDialogContent
-          css={{ maxWidth: 450, px: "$5", pt: "$4", pb: "$4" }}>
-          <AlertDialogTitle asChild>
-            <Heading size="1">
-              Delete{" "}
-              {state.selectedRows.length > 1 ? state.selectedRows.length : ""}{" "}
-              webhook
-              {state.selectedRows.length > 1 && "s"}?
-            </Heading>
-          </AlertDialogTitle>
-          <AlertDialogDescription asChild>
-            <Text
-              size="3"
-              variant="gray"
-              css={{ mt: "$2", lineHeight: "22px" }}>
-              This will permanently remove the webhook
-              {state.selectedRows.length > 1 && "s"}. This action cannot be
-              undone.
-            </Text>
-          </AlertDialogDescription>
+      <DeleteDialog
+        deleteDialogState={deleteDialogState}
+        state={state}
+        savingDeleteDialog={savingDeleteDialog}
+        setSavingDeleteDialog={setSavingDeleteDialog}
+        onDeleteWebhooks={onDeleteWebhooks}
+      />
 
-          <Flex css={{ jc: "flex-end", gap: "$3", mt: "$5" }}>
-            <AlertDialogCancel asChild>
-              <Button size="2" onClick={deleteDialogState.onOff} ghost>
-                Cancel
-              </Button>
-            </AlertDialogCancel>
-            <AlertDialogAction asChild>
-              <Button
-                size="2"
-                disabled={savingDeleteDialog}
-                onClick={async (e) => {
-                  try {
-                    e.preventDefault();
-                    setSavingDeleteDialog(true);
-                    await onDeleteWebhooks();
-                    openSnackbar(
-                      `${state.selectedRows.length} webhook${
-                        state.selectedRows.length > 1 ? "s" : ""
-                      } deleted.`
-                    );
-                    setSavingDeleteDialog(false);
-                    deleteDialogState.onOff();
-                  } catch (e) {
-                    setSavingDeleteDialog(false);
-                  }
-                }}
-                variant="red">
-                {savingDeleteDialog && (
-                  <Spinner
-                    css={{
-                      color: "$hiContrast",
-                      width: 16,
-                      height: 16,
-                      mr: "$2",
-                    }}
-                  />
-                )}
-                Delete
-              </Button>
-            </AlertDialogAction>
-          </Flex>
-        </AlertDialogContent>
-      </AlertDialog>
-
-      {/* Create webhook dialog */}
-      <WebhookDialog
+      <CreateDialog
         action={Action.Create}
         isOpen={createDialogState.on}
         onOpenChange={createDialogState.onToggle}
-        onSubmit={async ({ events, name, url }) => {
+        onSubmit={async ({ events, name, url, sharedSecret }) => {
           const newWebhook = await createWebhook({
             events,
             name,
             url,
+            sharedSecret,
           });
           await state.invalidate();
           const query = router.query.admin === "true" ? { admin: true } : {};
@@ -343,7 +114,7 @@ const WebhooksTable = ({ title = "Webhooks" }: { title?: string }) => {
           });
         }}
       />
-    </>
+    </Box>
   );
 };
