+++ packages/www/pages/dashboard/developers/webhooks/index.tsx
@@ -1,15 +1,15 @@
 import Layout from "layouts/dashboard";
 import { useApi, useLoggedIn } from "hooks";
-import { Box } from "@livepeer/design-system";
 import WebhooksTable from "components/Dashboard/WebhooksTable";
 
-const ApiKeys = () => {
+const Webhooks = () => {
   useLoggedIn();
   const { user } = useApi();
 
   if (!user) {
     return <Layout />;
   }
+
   return (
     <Layout
       id="developers/webhooks"
@@ -17,13 +17,9 @@ const ApiKeys = () => {
         { title: "Developers", href: "/dashboard/developers/webhooks" },
         { title: "Webhooks" },
       ]}>
-      <Box css={{ p: "$6" }}>
-        <Box css={{ mb: "$8" }}>
-          <WebhooksTable />
-        </Box>
-      </Box>
+      <WebhooksTable />
     </Layout>
   );
 };
 
-export default ApiKeys;
+export default Webhooks;
